refactor(authentication): move auth middleware to separate file
